Arm: Remove alternative from neon_softfp_fp16 directive.
There's a bit of a disconnect between the feature flags that don't test the fpu
and ones that do when the test itself also forces an architecture.  The forcing
of the architecture would change the defaults and without explicitly giving the
correct fpu again the test would fail.

I don't see a good way to solve this problem, really the feature tests should
ideally contain the extra options the test adds too, but for this specific case
it can be solved by always testing the fpu explicitly.

Committed under the GCC obvious

gcc/testsuite/ChangeLog:

	* lib/target-supports.exp
	(check_effective_target_arm_neon_softfp_fp16_ok_nocache): Drop non-fpu
	checking alternative.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@268943 138bc75d-0d04-0410-961f-82ee72b054a4