PR c++/88690 - C++17 ICE with empty base in aggregate.
Base fields for empty bases appear in initialization order, which may not be
the same as layout order.  If they also show up in a CONSTRUCTOR in that
order, output_constructor_regular_field aborts because it understandably
doesn't want to go backwards.  I also considered making o_c_r_f more
tolerant of the case where the out-of-order field has fieldsize 0, and so no
actual data needs to be emitted, but we might as well avoid adding an
element to the CONSTRUCTOR in the first place.

	* typeck2.c (process_init_constructor_record): Skip trivial
	initialization of an empty base.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@269073 138bc75d-0d04-0410-961f-82ee72b054a4