PR c++/86485 - -Wmaybe-unused with empty class ?:
The problem in this testcase is that the front end expects an rvalue
COND_EXPR to initialize a single temporary from one of the arms.  But
because gimplify_cond_expr used MODIFY_EXPR, instead the arms would each
create their own temporary and then copy that into the common temporary.

So, let's use INIT_EXPR instead.

	* gimplify.c (gimplify_cond_expr): Use INIT_EXPR.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@269403 138bc75d-0d04-0410-961f-82ee72b054a4