[PR86438] avoid too-long shift in test
The test fell back to long long and long when __int128 is not
available, but it assumed sizeof(long) < sizeof(long long) because of
a shift count that would be out of range for a long long if their
widths are the same.  Fixed by splitting it up into two shifts.


for  gcc/testsuite/ChangeLog

	PR rtl-optimization/86438
	* gcc.dg/torture/pr86438.c: Split up too-wide shift.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@270388 138bc75d-0d04-0410-961f-82ee72b054a4